+++ tests/providers/amazon/aws/operators/test_sagemaker_training.py
@@ -22,8 +22,8 @@ import unittest
 import mock
 
 from airflow.contrib.hooks.sagemaker_hook import SageMakerHook
-from airflow.contrib.operators.sagemaker_training_operator import SageMakerTrainingOperator
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.operators.sagemaker_training import SageMakerTrainingOperator
 
 role = 'arn:aws:iam:role/test-role'
